+++ Want2LearnBase/Program.cs
@@ -22,23 +22,7 @@ namespace Want2LearnBase
 
       
 
-        public static void BubbleSort(int[] Array)
-        {
-            int Temp;
-            for (int i = 0; i < Array.Length; i++)
-            {
-                for (int j = i + 1; j < Array.Length; j++)
-                {
-                    if (Array[i] > Array[j])
-                        {
-                        Temp = Array[i];
-                        Array[i] = Array[j];
-                        Array[j] = Temp;
-                        }
-                }
-            }
-           
-        }
+       
 
 
        
@@ -150,21 +134,7 @@ namespace Want2LearnBase
             return RandomString;
         }
 
-        public static char[] GetArrayOfEngRusLetters()
-        {
-            char[] EngAlphabet = GetEngAlphabet();
-            char[] RusAlphabet = GetRusAlphabet();
-            char[] AllLetters = new char[EngAlphabet.Length + RusAlphabet.Length];
-            for (int i = 0; i < EngAlphabet.Length; i++)
-            {
-                AllLetters[i] = EngAlphabet[i];
-            }
-            for (int i = EngAlphabet.Length, j = 0; i < AllLetters.Length; i++, j++)
-            {
-                AllLetters[i] = RusAlphabet[j];
-            }
-            return AllLetters;
-        }
+        
 
         public static string GetRandomWord(int stringLength, Random rnd)
         {
@@ -178,55 +148,7 @@ namespace Want2LearnBase
             return RandomString;
         }
 
-        public static void QSort(int[] array)
-        {
-            QSort(array, 0, array.Length - 1);
-        }
-
-        public static void QSort(int[] array, int left, int right)
-        {
-            if (left >= right)
-            {
-                return;
-            }
-            int pivot = left + (right - left) / 2;
-            Sorting(array, left, right, pivot);
-            QSort(array, left, pivot);
-            QSort(array, pivot + 1, right);
-            
-        }
-
-        public static void Sorting(int[] array, int left, int right, int pivot)
-        {
-            while (left < right)
-            {
-                if (array[left] < array[pivot])
-                {
-                    if (left < pivot)
-                    {
-                        left++;
-                    }
-                }
-                else
-                {
-                    if (array[right] > array[pivot])
-                    {
-                        if (right > pivot)
-                        {
-                            right--;
-                        }
-                    }
-                    else
-                    {
-                        Swap(ref array[left], ref array[right]);
-                        left++;
-                        right--;
-                    }
-                }
-                
-            }
-            
-        }
+       
 
 
     }
