+++ Want2LearnBase/Want2LearnBase.csproj
@@ -48,6 +48,7 @@
     <Compile Include="MatrixService.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SortService.cs" />
     <Compile Include="StringService.cs" />
     <Compile Include="SwapService.cs" />
   </ItemGroup>
